requirements.txt
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.59.2
fpdf2==2.8.4
openpyxl==3.1.5
pillow==11.3.0
pyodbc==5.2.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
python-dotenv==1.1.1